# docs/swagger.yaml
basePath: /
definitions:
  p2-ip-kartikajls_internal_dto.CreatePaymentRequest:
    properties:
      payment_method:
        type: string
      rental_order_id:
        type: integer
    required:
    - payment_method
    - rental_order_id
    type: object
...
  p2-ip-kartikajls_internal_dto.CreateTopUpRequest:
    properties:
      amount:
        type: number
    type: object
...
  p2-ip-kartikajls_internal_dto.UpdatePaymentStatusRequest:
    properties:
      payment_status:
        type: string
    required:
    - payment_status
    type: object
...
  p2-ip-kartikajls_internal_helper.Response:
    properties:
      responseCode:
        type: integer
      responseData: {}
      responseMessage:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API untuk aplikasi Rent Camera.
    API ini menyediakan fitur register, login, camera, top up, rental order, dan payment.
    Admin dapat melakukan approval terhadap top up, order, dan payment.
  title: Jati Rent Camera API
  version: "1.0"
paths:
...
  /admin/payments:
    get:
      description: Admin melihat seluruh pembayaran
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get All Payments
      tags:
      - Admin - Payment Approve
  /admin/payments/{payment_id}/status:
    put:
      consumes:
      - application/json
      description: Admin menyetujui pembayaran rental order
      parameters:
      - description: Payment ID
        format: int64
        in: path
        name: payment_id
        required: true
        type: integer
      - description: Payment Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.UpdatePaymentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Approve Payment
      tags:
      - Admin - Payment Approve
  /admin/topups:
    get:
      description: Admin melihat seluruh permintaan top up
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get All Top Ups
      tags:
      - Admin - Top Up Approve
  /admin/topups/{id}/approve:
    put:
      consumes:
      - application/json
      description: Admin menyetujui permintaan top up user
      parameters:
      - description: Top Up ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Approve Top Up
      tags:
      - Admin - Top Up Approve
...
  /users/payments:
    get:
      description: Mendapatkan seluruh pembayaran milik user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get My Payments
      tags:
      - User - Payment
    post:
      consumes:
      - application/json
      description: User melakukan pembayaran untuk rental order yang telah disetujui
      parameters:
      - description: Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Create Payment
      tags:
      - User - Payment
  /users/payments/{payment_id}:
    get:
      description: Mendapatkan detail pembayaran berdasarkan ID
      parameters:
      - description: Payment ID
        format: int64
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get My Payment By ID
      tags:
      - User - Payment
...
  /users/topup/{id}:
    get:
      description: Mendapatkan detail top up berdasarkan ID
      parameters:
      - description: Top Up ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get Top Up By ID
      tags:
      - User - Top Up
  /users/topups:
    get:
      consumes:
      - application/json
      description: User melihat riwayat top up miliknya
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get My Top Ups
      tags:
      - User - Top Up
    post:
      consumes:
      - application/json
      description: User membuat permintaan top up saldo
      parameters:
      - description: Top Up Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.CreateTopUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Create Top Up
      tags:
      - User - Top Up
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"